[settings] Remove base::NoDestructor use from FilesSection
The arrays of SearchConcepts are static data and do not need to be
stored in a std::vector. One tricky bit is that these arrays were
initialized based on the result of checking a base::Feature; luckily the
base::Feature has been launched for some time and the feature check can
simply be removed.

Bug: 370837212
Bug: 375364966
Change-Id: I2c09398f950590cea2af49ddbb6278500a25c52b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5984249
Reviewed-by: Austin Tankiang <austinct@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1377157}