[PWA/NavCapture] Include query string in expectations, and preserve on redirect
Include the query string in the expectations file to make it easier to
see what URLs are actually involved in various places. Additionally
preserve the query string on redirects. This helps to ensure that the
destination URL in more cases won't be identical to the "start" url of
the installed apps, as them matching causes some edge cases where we
might end up navigating to the same URL we are already at, which behaves
differently.

Bug: 351775835
Change-Id: Ie21c38b66b4ae9280c02b7c515b255d78333b7f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5983983
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1377167}