[settings] Remove NoDestructor use from AppsSection
The arrays of SearchConcepts are static data and do not need to be
stored in a std::vector. One tricky bit is that these arrays were
initialized based on the result of checking a base::Feature; luckily the
base::Feature has been launched for some time and the feature check can
simply be removed.

Bug: 370837212
Bug: 375364966
Change-Id: I84c4cd1b4f8ba92b33068da82fbc1f74e0dbfd67
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5984244
Reviewed-by: Wes Okuhara <wesokuhara@google.com>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1377171}