[PDF Ink Signatures] Add skeleton pdfium_ink_reader code
Add pdf/pdfium/pdfium_ink_reader.h, which contains a function named
ReadV2InkPathsFromPageAsModeledShapes() to import "V2" Ink paths from a
FPDF_PAGE object. The function has a skeleton implementation, which adds
structure, but effectively does nothing. Add unit tests for this new
function, which is similarly incomplete as there is not much to test for
now.

The newly added ink_v2.pdf is generated as follows:

- Open pdf/test/data/blank.pdf in an existing Ink application.
- Doodle on the page.
- Save PDF.

As the PDF is machine-generated and provided as-is to be authentic, it
has no corresponding .in file.

Bug: 353942910
Change-Id: I413154c6bd29bfb7268509a203731ca33ffed205
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5979237
Reviewed-by: Alan Screen <awscreen@chromium.org>
Reviewed-by: Andy Phan <andyphan@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1377179}