[HFM] Ignore Typically Secure fallbacks for non-unique hostnames
Presently, if a user navigates to a non-unique hostname, the Typically
Secure heuristic will count this as a fallback event. The heuristic
can only enable HFM in Balanced Mode which doesn't upgrade non-unique
hostnames, so there's a mismatch in behavior.

This CL stops recording navigations to non-unique hostnames as fallback
events for the Typically Secure heuristic. This will result in a
slightly larger number of users to be opted into HFM due to the
heuristic. For example, users who navigated to their routers recently
would not have HFM auto-enabled before, but they will have it
auto-enabled if all other requirements of the Typically Secure
heuristic is satisfied.

The CL also rewrites the browser test for non-unique hostnames for the
heuristic. The previous test didn't work correctly: It was testing
that non-unique hostnames didn't show an interstitial in Typically
Secure mode, but the conditions of the heuristic weren't satisfied
either, so the test was a no-op.

Bug: 371000390
Change-Id: I06a0108a0bce56c53c2f86e3633128586d4cbab8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5967338
Reviewed-by: Chris Thompson <cthomp@chromium.org>
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1377182}