[settings] Remove base::NoDestructor use from MultitaskingSection
The arrays of SearchConcepts are static data and do not need to be
stored in a std::vector. One tricky bit is that these arrays were
initialized based on the result of checking a base::Feature; luckily the
base::Feature has been launched for some time and the feature check can
simply be removed.

Bug: 370837212
Bug: 375364966
Change-Id: Ib4a76f25902cd6f7d36e0e75130e4b33ed1521d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5984253
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Wes Okuhara <wesokuhara@google.com>
Cr-Commit-Position: refs/heads/main@{#1377183}