[PDF Ink Signatures] Add size selector dropdown in bottom toolbar
Add a size selector dropdown menu and button so that users can choose
brush sizes. Clicking the button will show the dropdown menu with
brush size selection.

To do so, introduce ViewerBottomToolbarDropdownElement to have a
button that opens a dropdown menu. This can be re-used when the color
selector is added in the future.

Low-Coverage-Reason: COVERAGE_UNDERREPORTED This CL adds tests that use
the bottom toolbar element and the new bottom toolbar dropdown element.

Bug: 369653190
Change-Id: I53722e395eb6e67a9603f000a301fe34131d9e4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5980052
Commit-Queue: Andy Phan <andyphan@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1377189}