Apply auto wide display logic to SettingsActivity
On wide displays, certain activities require some custom padding to fit their content to a reasonable width. These calculations assume that Chrome extends across the entire width of the screen, so by default they work with the window screen width. This causes issues on some automotive devices, where part of the screen is reserved for other UI, and Chrome is only given a limited portion, leading to higher padding values that "squeeze" the content on those activities.

This change applies the fix used for the NTP to the Settings activity, where the base view's measured width is used instead of the configuration's screen width. This logic could apply to any device, but is gated to auto devices to narrow scope.

Bug: 305300053
Change-Id: I6a04a24a77f495ecb9696ae2eec35efb2271d740
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5985357
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Theresa Sullivan <twellington@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1377208}