[settings] Remove base::NoDestructor use from PrintingSection
The arrays of SearchConcepts are static data and do not need to be
stored in a std::vector. One tricky bit is that these arrays were
initialized based on the result of checking a base::Feature; luckily the
base::Feature has been launched for some time and the feature check can
simply be removed.

Bug: 370837605
Bug: 375364966
Change-Id: I224fea082e488300b2031290ed8f6a13aed2c394
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5984257
Reviewed-by: Wes Okuhara <wesokuhara@google.com>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1377217}