[ash] speculative fix for crash in quick settings
QuickSettingsMediaViewController holds an untriaged raw ptr to
the most recent QuickSettingsMediaView that it creates, along
with one use that does a null check since (according to the
comment) the view can be destroyed first sometimes.  Unfortunately,
nothing seems to clear the raw ptr.

This CL replaces the rawptr with a weak ptr.  Since ownership of
the view isn't obviously tied to the controller in any way, and
since the view also holds a (triaged) rawptr back to the controller,
it seems likely that the controller => view pointer is sometimes
dangling.  Whether this fixes this particular crash is not known.

Bug: 368135848
Change-Id: I2a65915051e03ccb06899a5badd8270be5710d47
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5985274
Commit-Queue: Frank Liberato <liberato@chromium.org>
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1377253}