Log touchscreen gesture events in queue for debugging.
This CL adds logging to investigate the root cause of the associated
issue. It retains important data about TouchscreenGestureEvent
sequences, so that when they collide with a TouchpadGestureEvent
sequence that is scroll bubbling, it may be possible to understand how
the event router got into the bad state.

Bug: 346629231
Change-Id: I675c83f3911dc52cb9a7cdd4efb3c15a8e30ecfe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5979009
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Commit-Queue: James Maclean <wjmaclean@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1377400}