scanner: Allow ActionButtonView to collapse into an IconButton.
The UX specs state that a collapsed action button should be an
IconButton instead of a PillButton.

Currently, ActionButtonView is just a PillButton. To allow it to
collapse into an IconButton, convert ActionButtonView into a plain
Button that shows both text and icon by default, but can be collapsed
into just an icon.

Bug: b:374356291
Change-Id: I128a1614a018f2577d665e2e4e3a86423c7d8a67
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5981890
Reviewed-by: Sophie Wen <sophiewen@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: Michelle Chen <michellegc@google.com>
Cr-Commit-Position: refs/heads/main@{#1377460}