scanner: Add consent disclaimer dialog
The consent dialog allows users to accept or decline the sunfish
feature's terms and conditions.

Clicking accept will record the acceptance in pref and the user will be
able to use the feature.

If declined, we won't enter sunfish mode. Next time the user clicks on
the sunfish button, the window will pop up again asking to accept.

The disclaimer appears over the capture mode overlay and disables mouse
events in the capture mode session while it is visible.

In the future when we implement the ability to toggle this feature in
the setting, declining should disable the feature entirely in the
setting using a different pref.

UI layout logic is based on
https://source.chromium.org/chromium/chromium/src/+/2f7fefe7b5cb5147e1c09455046309f1b7e1188a:ash/system/magic_boost/magic_boost_disclaimer_view.h;bpv=1;bpt=0

UI screenshot is here: go/screenshots/6H86qcyYYZE64to

Bug: b:367882127
Change-Id: I3c4c5cd7ac11d57b60a8fe6947dba5b550a089d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5981208
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Auto-Submit: Jon Htin <jhtin@google.com>
Reviewed-by: Sophie Wen <sophiewen@chromium.org>
Commit-Queue: Jon Htin <jhtin@google.com>
Cr-Commit-Position: refs/heads/main@{#1377461}