//ash: Gate ViewTreeHostRootViewFrameFactory SCANOUT usage on SI support
ViewTreeHostRootViewFrameFactory adds SCANOUT SharedImage usage in
contexts where overlays are not actually supported (e.g., Flex in
production and ChromeOS-on-Linux in testing). As a result,
GLTextureImageBacking needs to support SCANOUT usage even though it is
not actually capable of producing overlays.

This CL guards ViewTreeHostRootViewFrameFactory's addition of SCANOUT
usage by the relevant SharedImage support for scanout being present.
For safety, this change is under a killswitch.

Bug: 330865436
Change-Id: Ib4291e4e21a236813b36d3e07f5f180d86d84df9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5980031
Reviewed-by: Zoraiz Naeem <zoraiznaeem@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1377549}