[iOS] Feature kSeparateProfilesForManagedAccounts: Check min iOS version
This feature depends on multi-profile infrastructure, which is only
available on iOS 17+.
So this CL adds a helper AreSeparateProfilesForManagedAccountsEnabled()
to check that the feature flag is enabled and the required iOS version
is available.

Bug: 331783685
Change-Id: Ie727475f158e3b5c158bc6d6cc0f6aab6cad04bd
Low-Coverage-Reason: HARD_TO_TEST The uncovered lines are for iOS <= 16 which apparently isn't covered by bots anymore. They're also entirely trivial, just an early-out into the feature-disabled code path.
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5980213
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1377562}