Remove some prefill metric
As discussed offline, the metric was used to evaluate a feature that
already launched, and there's no plan to analyze anything given by this
metric in the near future.

This simplifies the filling logic a little, which will be useful in
order to reach state where we have a single filling function instead of
three.

Bug: 377228667
Change-Id: I5e3b29276aab73e23e30ee98eacf79a1874bfa6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5987910
Commit-Queue: Jihad Hanna <jihadghanna@google.com>
Reviewed-by: Dominic Battré <battre@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1377579}