[High5] Do not show PinSetupScreen twice
Do not show the PinSetupScreen a second time when the user chooses to
have a password instead. This introduces a new PinSetupMode that will
be used for this purpose. Due to the new setup mode, a boolean that
was being used for controlling the visibility of the back button on the
PasswordSelectionScreen becomes redundant and was thus removed.

Bug: b/365059362, b/348326316, b/374882602, b/375592778
Change-Id: I0348d7c6539682834772563772fea6d9697429a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5975563
Reviewed-by: Simon Hangl <simonha@google.com>
Commit-Queue: Renato Silva <rrsilva@google.com>
Reviewed-by: Istvan Nagy <iscsi@google.com>
Reviewed-by: Denis Kuznetsov <antrim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1377590}