[iOS] Run AccountProfileMapper unit tests only on iOS 17+
The underlying functionality is only available on iOS 17+, and as of
crrev.com/c/5980213, the feature is only considered enabled on iOS 17+.
This broke the unit tests on iOS 16 bots.

This CL skips the affected tests on iOS 16, and adds a test to verify
the feature state is as expected.

Bug: 331783685
Change-Id: Ic5f9c07b9278e98f172f5cc3f9fdd5243850b546
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5988411
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1377624}