[ios] Prototype API to change the profile for a scene
Add a new command to change the profile for a scene. The API is
asynchronous and invokes the completion on success/failure. For
success the completion is called when the profile initialisation
is complete (in case the profile is not yet initialised).

There is no animation for the transition, and the API does not
really provides a way to implement an animation (to do that, it
should probably be changed to take a delegate that is called with
an UIViewController* to use as the root for the animation, but it
will require crbug.com/353683675 to be complete first).

Bug: 353684413
Low-Coverage-Reason: difficult to test and experimental.
Change-Id: I24c48d4b8fb944438fd6c63507abcfa111a91106
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5980028
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Aliona Dangla <alionadangla@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1377630}