[ios] Autofill across iframes: add switch to disable triggered fetches
This switch is enabled by default and allows disabling the
form scan triggered when the forms forest calls
TriggerFormExtractionInDriverFrame().

TriggerFormExtractionInDriverFrame() can be expensive on iOS, we're
working on optimizing the cost of form fetches through throtttling
but this may take some time to deploy, so in the meantime we thought
that disabling the scans could help launching xframe in the case
terminations are judged problematic.

The switch keeps the status quo by default.

Change-Id: I2ac25d925891a00bd86f9006dd1ccae9790729d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5987311
Commit-Queue: Vincent Boisselle <vincb@google.com>
Reviewed-by: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#1377664}