[Android] Fix crash in MFControllerImpl::OnSourceAvailabilityChanged.
MF = manual filling

MFControllerImpl fetches sheet data from the individual sheet controller
when the list of suggestions should be updated. The sheet data is
represented as optional. This CL adds a check to make sure the optional
data is present before sending it to the Android UI.

Bug: 375224895
Change-Id: I74cdc3a3b2cf647444401a198382f46cc33b4cea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5982827
Reviewed-by: Slobodan Pejic <slobodan@chromium.org>
Commit-Queue: Timofey Chudakov <tchudakov@google.com>
Cr-Commit-Position: refs/heads/main@{#1377671}