Fix blink web tests "--target" arg in wrapper scripts
The "name" arg to get_path_info() assumes the full filepath is a file
and not a dir. Consequently, it chops off the trailing "." from any
path. So when passing a build-dir path to it like
"//out/some.build.dir", it will return "some.build" instead of
"some.build.dir" that we want. This is the source of failures like:
https://ci.chromium.org/ui/p/chromium/builders/try/mac12.0-blink-rel/7687/infra

Using "file" though will behave how we want, ie: just giving us the
last dir name of the build dir path. It's a little weird to passing
"file" when we really want the dir name. But it seems to work.

And roll the per-builder-build-dir experiment on blink trybots forward
again. This effectively revert https://crrev.com/c/5983847 since it
should be safe now with the GN fix.

Led builds with this CL and the experiment restored:
https://ci.chromium.org/b/8732452888242765697
https://ci.chromium.org/b/8732452867352051937

Bug: 355218109
Change-Id: Ic884717cc5f0756128d20d15306ef08c9fbe8344
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5980565
Reviewed-by: Dirk Pranke <dpranke@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Weizhong Xia <weizhong@google.com>
Cr-Commit-Position: refs/heads/main@{#1377704}