build: use --target=*-cros-* when building for ChromeOS devices
We used to pass generic target triples when building for ChromeOS
devices. This doesn't accurately reflect the target being built for, and
also doesn't work in all cases. For example, the bundled LLVM toolchain
does not have libraries for --target=arm-linux-gnueabihf, which is
what we would pass when targeting 32-bit ARM.

This change simplifies some of the logic for determining the target
triple to pass so that when we are building for a ChromeOS device
(is_chromeos_device is true), we use one of the *-cros-* target
triples instead of the generic ones.

For the case of 32-bit ARM, this changes the target triple from
arm-linux-gnueabihf (which we don't bundle libraries for)
to armv7a-cros-linux-gnueabihf (which we do bundle libraries for).

Bug: b:376446789
Test: emerge-${BOARD} chromeos-chrome
Change-Id: I382bf54fc627f91c0aba8578c86acba733502631
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5984733
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Bob Haarman <inglorion@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1377710}