Uprev ChromeScreenAI CIPD to 126.18
This CL updates the CIPD for downloading the library that is used for
ChromeSreenAI browser tests.

If browser tests fail, revert the rollout for the new binary in the
component updater as well. You will need to rollout a newer (or older)
release to update the clients that are affected by the faulty rollout.

Change-Id: Ifb39a7ea835238ea4706a6b38610133838f3644a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5985234
Reviewed-by: Kyungjun Lee <kyungjunlee@google.com>
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1377727}