[DNT] Aborting the animation when the renderer crashes
Added a new function override from `WebContentsObserver` called `PrimaryMainFrameRenderProcessGone` which gets notified if the primary main frame renderer is destroyed. In that case, the animation manager destroys the animator when the function is called. Also, added a new enum value to keep track of this animation abort reason in the UMA metrics.

Bug: 328601834
Change-Id: I009ba9d4cea1ec17a69a0fe81a1879ce4bf4d050
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5979352
Reviewed-by: William Liu <liuwilliam@chromium.org>
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Commit-Queue: Baran Erfani <baranerf@google.com>
Cr-Commit-Position: refs/heads/main@{#1377752}