Add testing config for HttpsFirstBalancedModeAutoEnable
This enables the kHttpsFirstBalancedModeAutoEnable feature flag in tests
and fixes the last set tests affected by this change. Currently this is
only enabled on Desktop platforms.

For HttpsUpgradesBrowsertest this removes BalancedModeByDefault from the
kAllAutoHFM set, since it is already covered by the
kHttpsFirstBalancedMode set. For TypedNavigationUpgrades this disabled
the auto-enable feature entirely.

Due to the large number of affected files, this change is being split into multiple CLS -- see crrev.com/c/5787951/19 for the full set of
affected files. The other test fix CLs that were split off are:

- crrev.com/c/5979931
- crrev.com/c/5982268
- crrev.com/c/5982269
- crrev.com/c/5980847
- crrev.com/c/5980794
- crrev.com/c/5980259
- crrev.com/c/5982289
- crrev.com/c/5983018
- crrev.com/c/5980877
- crrev.com/c/5982190
- crrev.com/c/5983203
- crrev.com/c/5979603
- crrev.com/c/5980857
- crrev.com/c/5983224
- crrev.com/c/5980413
- crrev.com/c/5979604
- crrev.com/c/5983204
- crrev.com/c/5979605
- crrev.com/c/5983205
- crrev.com/c/5981831

Bug: 349860796
Change-Id: Iec9968113b0f7dfc3438b3c38becd9eccfc33097
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5983207
Commit-Queue: Chris Thompson <cthomp@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1377847}