[PDF Ink Signatures] Remove TODOs in UndoRedoInvalidations test cases
After checking the values manually, they are in fact correct. This is
because InkStrokeModeler takes the sharp test inputs and smoothed them
out. See comments on the bug for details. Replace the TODOs with
comments that explain why the test expectations are correct as-is.

Bug: 376301209
Change-Id: I986f61255730d05db61b1c82230317b42fc166d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5979614
Reviewed-by: Alan Screen <awscreen@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1377869}