Fix CCT startup metrics for early navigation
EarlyNav was mistakenly getting grouped with speculative prerenders.
Instead, track the metrics for the navigation as soon as it starts (as
it's not speculative).

The added isHidden code in
https://chromium-review.googlesource.com/c/chromium/src/+/5936843
to not record metrics when the tab is hidden was actually unnecessary,
that was already covered by the fact that we don't record metrics until
the launch timestamps are sent to the observer, which for hidden tabs
is after the Activity starts.

Bug: 372710946
Change-Id: I0999a9c650263f8f911813cb286a4615373820d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5989053
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1377912}