[SAA] Fix local-/session-storage access via handle w/ 3p cookie access
The fix in:
https://chromium-review.googlesource.com/c/chromium/src/+/5675704
ensured SAA handle access when 3PC were on or the permission was
granted. This is important as if 3PC are on, the handle will always be
possible to generate.

The prior fix failed to extend this same support to
local-/session-storage access, which should have been done as these
interfaces are bound outside of the SAA handle so need their own check.

The checks are now merged to prevent future issues and new tests were
added.

Bug: 370937463
Change-Id: I712c0ea136f62a378ffa2b4647898165f8aee579
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5988711
Auto-Submit: Ari Chivukula <arichiv@chromium.org>
Commit-Queue: Ari Chivukula <arichiv@chromium.org>
Reviewed-by: Chris Fredrickson <cfredric@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1377949}